refactor(discord): move runtime restored state helpers
